// server/src/modules/statistic/dto/add-statistic.dto.ts
export class AddStatisticDto {
  employeeId: number;

  projectId: number;

  countMessage: number;

  countCharInMessage: number;

  countCommits: number;

  countChangedCode: number;

  countHoursInApps: number;

  countTasks: number;

  countDaysOnTask: number;

  lastUpGradeDate: number;

  countWorkYearInCompany: number;

  countCall: number;

  callDuration: number;

  countWorkHours: number;

  countWorkHoursOver: number;

  countDaysSickTime: number;

  countMissDeadline: number;

  countOutstandingTask: number;
}